+++ packages/brick_json_generators/lib/json_deserialize.dart
@@ -66,8 +66,15 @@ mixin JsonDeserialize<_Model extends Model, _Annotation extends FieldSerializabl
 
       // Iterable<enum>
       if (argTypeChecker.isEnum) {
+        final deserializeFactory = argTypeChecker.enumDeserializeFactory(providerName);
+        if (deserializeFactory != null) {
+          final nullableSuffix = argTypeChecker.isNullable ? '?' : '';
+          return '$fieldValue$nullableSuffix.map(${SharedChecker.withoutNullability(argType)}.$deserializeFactory)';
+        }
+
         if (fieldAnnotation.enumAsString) {
           final nullableSuffix = argTypeChecker.isNullable ? '' : '!';
+
           return '''$fieldValue.map(
             (value) => ${providerName}Adapter.enumValueFromName(${SharedChecker.withoutNullability(argType)}.values, value)$nullableSuffix
           )$castIterable$defaultValue
@@ -106,6 +113,11 @@ mixin JsonDeserialize<_Model extends Model, _Annotation extends FieldSerializabl
 
       // enum
     } else if (checker.isEnum) {
+      final deserializeFactory = checker.enumDeserializeFactory(providerName);
+      if (deserializeFactory != null) {
+        return '${checker.isNullable ? "$fieldValue == null ? null :" : ""} ${SharedChecker.withoutNullability(field.type)}.$deserializeFactory($fieldValue)';
+      }
+
       if (fieldAnnotation.enumAsString) {
         final nullableSuffix = checker.isNullable ? '' : '!';
         return '${providerName}Adapter.enumValueFromName(${SharedChecker.withoutNullability(field.type)}.values, $fieldValue)$nullableSuffix$defaultValue';