+++ packages/brick_json_generators/CHANGELOG.md
@@ -1,5 +1,9 @@
 ## Unreleased
 
+## 1.1.1
+
+* Respect enum `from<Provider>` constructors and `to<Provider>` methods
+
 ## 1.1.0
 
 * Upgrade analyzer to version 4