+++ packages/brick_offline_first_build/CHANGELOG.md
@@ -1,5 +1,9 @@
 ## Unreleased
 
+## 2.1.1
+
+* Respect enum `from<Provider>` constructors and `to<Provider>` methods
+
 ## 2.1.0
 
 * Upgrade analyzer to version 4